+++ tests/testthat/test-adjust_p.R
@@ -9,7 +9,7 @@ test_that("adjust_p return the correct class", {
     tidy_summary(prtest = "P") %>%
     adjust_p()
 
-  expect_is(adj, "tidy_summary")
+  expect_s3_class(adj, "tidy_summary")
 })
 
 test_that("warning and return if test is not set", {
@@ -24,5 +24,5 @@ test_that("warning and return if test is not set", {
       adjust_p()
   )
 
-  expect_is(no_test_adj, "tidy_summary")
+  expect_s3_class(no_test_adj, "tidy_summary")
 })