+++ tests/testthat/test-check_for_bot_options.R
@@ -1,5 +1,3 @@
-context("test-check_for_bot_options")
-
 test_that("return correct output", {
   expect_equal(suppressWarnings(check_for_bot_options()), FALSE)
 
@@ -18,7 +16,7 @@ test_that("return correct output", {
 
 
 test_that("return a Warning if false", {
-  expect_warning(check_for_bot_options())
+  expect_warning(expect_warning(check_for_bot_options()))
   expect_warning(check_for_bot_options(bot = 1))
   expect_warning(check_for_bot_options(chat_id = 1))
 })