+++ tests/testthat/test-check_pkg.R
@@ -1,5 +1,3 @@
-context("test-install_pkg_set")
-
 test_that("empty please_install()", {
 
   expect_null(please_install(character(0)))