+++ tests/testthat/test-ci2p.R
@@ -1,5 +1,3 @@
-context("test-ci2p")
-
 test_that("known results", {
   expect_equal(
     ci2p(1.125, 0.634, 1.999, log_transform = TRUE),
@@ -7,7 +5,8 @@ test_that("known results", {
   )
   expect_equal(
     ci2p(1.257, 1.126, 1.403, log_transform = TRUE),
-    0.0003118063
+    0.0003118063,
+    tolerance = 1e-7
   )
 })
 
@@ -30,6 +29,7 @@ test_that("inverted lower and upper", {
     suppressWarnings(
       ci2p(1.257, 1.403, 1.126, log_transform = TRUE)
     ),
-    0.0003118063
+    0.0003118063,
+    tolerance = 1e-7
   )
 })